# docs/swagger.yaml
definitions:
...
  controllers.RatingInput:
    properties:
      description:
        type: string
      productId:
        type: integer
      star:
        type: integer
    type: object
...
  models.Review:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      productId:
        type: integer
      star:
        type: integer
      updatedAt:
        type: string
      usersId:
        type: integer
    type: object
  user.ChangePassword:
    properties:
      email:
        type: string
      passwordNew:
        type: string
      passwordOld:
        type: string
    required:
    - email
    - passwordNew
    - passwordOld
    type: object
  user.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.RegisterUserInput:
    properties:
      dateOfBirth:
        type: string
      email:
        type: string
      gender:
        type: string
      name:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  user.UserFormatter:
    properties:
      Token:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
      role:
        type: string
      string:
        type: string
    type: object
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /changePassword:
    post:
      description: Change Password User.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Change Password an user.
      tags:
      - Auth
...
  /login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserFormatter'
      summary: Login as as user.
      tags:
      - Auth
...
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserFormatter'
      summary: Register a user.
      tags:
      - Auth
  /review:
    get:
      description: Get list Review refrence By userID
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Get Review By User Id
      tags:
      - Review
    post:
      description: Creating a new Review.
      parameters:
      - description: the body to create a new Review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RatingInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Create New Review.
      tags:
      - Review
  /review/{id}:
    delete:
      description: Delete a Review by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Review.
      tags:
      - Review
    patch:
      description: Update Review by id.
      parameters:
      - description: the body to update a new Review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RatingInput'
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Update Review.
      tags:
      - Review
swagger: "2.0"